Deprecate PolisClient as legacy.
